Create the responsive search & options section
This "looks" like the reference design I made, but most of this will
NOT be functional, as it's not necessary for the specs required of the
library project from TOP. To do it right the first time, would be a
waste of the time relative to what I actually have to do, and I don't
think I will actually implement the functionality those design parts
require. Maybe I will, but it'd likely require custom select controls,
which I desperately would want to avoid doing for the sake of sanity &
accessibility.

I did use z-index on the header, because to position the search SVG on
the input cleanly, I used "position: relative" on its wrapper, and that
led to that content being visible behind the header on scroll, even if
on page-load there was no issues of content displaying as intended.
I only used z-index, because there was no other solution I could find.

But hey, it looks like my reference design, and is quite responsive.