Create the responsive catalog card
I'm aware that this should be generated progammatically in the future.
But I chose to still do this so I had an effective idea of precisely
how I want it to be positioned, look, and thus generate accordingly.

There's also a point where the mark read & completed icons cause
undesired behavior right now, but those will be binary icon options
when functionality is implemented, so it's actually fine.